Perform the nmstate network configuration using the DHCP too
EIB generated images can consume specific network configurations
provided by CAPI, when the "configure-network.sh" script is located in
the network directory at build time.

This script is required to parse the nm-state formatted network
configuration and has been historically not included for hosts that
would be provisioned through DHCP.

However, while DHCP will work in absence of such script, no additional
configuration is possible. This can be particularly useful when trying to
fine-tune the autoconfiguration for IPv6, or when network
autoconfiguration is desired but some values (e.g. DNSes) must be
statically provided.

This commit includes the "configure-network.sh" script in the EIB
examples for DHCP.

Signed-off-by: Marco Chiappero <marco.chiappero@suse.com>